requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in requirements.in
#
altair==5.5.0
    # via streamlit
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via starlette
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.1
    # via streamlit
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via
    #   reportlab
    #   requests
click==8.3.0
    # via
    #   nltk
    #   streamlit
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   loguru
    #   pytest
    #   tqdm
easyocr==1.7.2
    # via -r requirements.in
faiss-cpu==1.12.0
    # via -r requirements.in
fastapi==0.120.2
    # via -r requirements.in
filelock==3.20.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.9.0
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
h11==0.16.0
    # via uvicorn
huggingface-hub==0.36.0
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   requests
imageio==2.37.0
    # via scikit-image
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.5.2
    # via
    #   nltk
    #   scikit-learn
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
keybert==0.9.0
    # via -r requirements.in
lazy-loader==0.4
    # via scikit-image
loguru==0.7.3
    # via -r requirements.in
lxml==6.0.2
    # via python-docx
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
narwhals==2.10.0
    # via altair
networkx==3.5
    # via
    #   scikit-image
    #   torch
ninja==1.13.0
    # via easyocr
nltk==3.9.2
    # via -r requirements.in
numpy==2.2.6
    # via
    #   -r requirements.in
    #   easyocr
    #   faiss-cpu
    #   imageio
    #   keybert
    #   opencv-python
    #   opencv-python-headless
    #   pandas
    #   pydeck
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   shapely
    #   streamlit
    #   tifffile
    #   torchvision
    #   transformers
opencv-python==4.12.0.88
    # via -r requirements.in
opencv-python-headless==4.12.0.88
    # via easyocr
packaging==25.0
    # via
    #   altair
    #   faiss-cpu
    #   huggingface-hub
    #   lazy-loader
    #   pytest
    #   scikit-image
    #   streamlit
    #   transformers
pandas==2.3.3
    # via
    #   -r requirements.in
    #   streamlit
pillow==12.0.0
    # via
    #   easyocr
    #   imageio
    #   reportlab
    #   scikit-image
    #   sentence-transformers
    #   streamlit
    #   torchvision
pluggy==1.6.0
    # via pytest
protobuf==6.33.0
    # via streamlit
pyarrow==21.0.0
    # via streamlit
pyclipper==1.3.0.post6
    # via easyocr
pydantic==2.12.3
    # via fastapi
pydantic-core==2.41.4
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via
    #   pytest
    #   rich
pymupdf==1.26.5
    # via -r requirements.in
pypdf==6.1.3
    # via -r requirements.in
pytest==8.4.2
    # via -r requirements.in
python-bidi==0.6.7
    # via easyocr
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.2.0
    # via -r requirements.in
python-dotenv==1.2.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   easyocr
    #   huggingface-hub
    #   transformers
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.10.23
    # via
    #   nltk
    #   transformers
reportlab==4.4.4
    # via -r requirements.in
requests==2.32.5
    # via
    #   huggingface-hub
    #   streamlit
    #   transformers
rich==14.2.0
    # via keybert
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
safetensors==0.6.2
    # via transformers
scikit-image==0.25.2
    # via easyocr
scikit-learn==1.7.2
    # via
    #   -r requirements.in
    #   keybert
    #   sentence-transformers
scipy==1.16.3
    # via
    #   easyocr
    #   scikit-image
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.1.2
    # via
    #   -r requirements.in
    #   keybert
shapely==2.1.2
    # via easyocr
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
starlette==0.49.1
    # via fastapi
streamlit==1.51.0
    # via -r requirements.in
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via streamlit
threadpoolctl==3.6.0
    # via scikit-learn
tifffile==2025.10.16
    # via scikit-image
tokenizers==0.22.1
    # via transformers
toml==0.10.2
    # via streamlit
torch==2.9.0
    # via
    #   easyocr
    #   sentence-transformers
    #   torchvision
torchvision==0.24.0
    # via easyocr
tornado==6.5.2
    # via streamlit
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   nltk
    #   sentence-transformers
    #   transformers
transformers==4.57.1
    # via sentence-transformers
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   referencing
    #   sentence-transformers
    #   starlette
    #   streamlit
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via -r requirements.in
watchdog==6.0.0
    # via streamlit
win32-setctime==1.2.0
    # via loguru

# The following packages are considered to be unsafe in a requirements file:
# setuptools
